refactor: move pre_release to coordinators
