+++ app/libs/coordinators.rb
@@ -43,7 +43,7 @@ module Coordinators
     def self.release_has_started!(release)
       Coordinators::SetupReleaseSpecificChannel.call(release)
       release.notify!("New release has commenced!", :release_started, release.notification_params)
-      Releases::PreReleaseJob.perform_async(release.id)
+      Coordinators::PreReleaseJob.perform_async(release.id)
       Releases::FetchCommitLogJob.perform_async(release.id)
       RefreshReportsJob.perform_async(release.hotfixed_from.id) if release.hotfix?
     end
